Fix compiler warnings for sign conversion and deprecated function prototypes
- Add explicit casts for int/short to unsigned int in X11 API calls:
  * callbacks.c: XResizeWindow width/height parameters
  * catalog.c: XCopyArea coordinate and dimension parameters

- Fix incompatible pointer types in catalog.c:
  * Use temporary int variables when calling catalogFindOpenLocation
  * which expects int* but struct fields are short*

- Update function pointer types in cexec.c:
  * Replace deprecated () prototype syntax with explicit types
  * Use proper VObj*, Packet*, int, Packet*, long signatures
  * from class.h for slotGetMeth and slotSetMeth

- Modernize freePossibleDangler function:
  * Add proper ANSI C prototype in packet.h
  * Convert K&R function definition to ANSI C in packet.c

All fixes maintain runtime behavior while ensuring C23 compatibility.

+++ src/vw/callbacks.c
@@ -902,7 +902,7 @@ void resizeViola(Widget widget, XtPointer clientData, XEvent* event, Boolean* co
     if (event->type == ConfigureNotify) {
         XConfigureEvent* xcep = (XConfigureEvent*)event;
 
-        XResizeWindow(XtDisplay(widget), (Window)clientData, xcep->width, xcep->height);
+        XResizeWindow(XtDisplay(widget), (Window)clientData, (unsigned int)xcep->width, (unsigned int)xcep->height);
     }
 }
